Make button Share in Counter panel child element.
